Removed openGL pipeline and reworked SDL pipeline to not use GLSL shaders
