+++ src/renderer/renderer.cpp
@@ -1,0 +1,82 @@
+#include <SDL3/SDL.h>
+
+#include <main.h>
+
+extern "C" {
+// ---------------------------------------------------------------------------
+// SDL3 app callbacks
+// ---------------------------------------------------------------------------
+
+SDL_AppResult SDL_AppInit(void** appstate, int /*argc*/, char* /*argv*/[])
+{
+    SDL_SetAppMetadata("group2", "0.1.0", "com.cse125.group2");
+
+    if (!SDL_Init(SDL_INIT_VIDEO)) {
+        SDL_Log("SDL_Init failed: %s", SDL_GetError());
+        return SDL_APP_FAILURE;
+    }
+
+    auto* s = new AppState();
+    *appstate = s;
+
+    constexpr int k_winW = 1280;
+    constexpr int k_winH = 720;
+    constexpr SDL_WindowFlags k_winFlags = SDL_WINDOW_RESIZABLE;
+
+    s->window = SDL_CreateWindow("group2", k_winW, k_winH, k_winFlags);
+    if (!s->window) {
+        SDL_Log("SDL_CreateWindow failed: %s", SDL_GetError());
+        return SDL_APP_FAILURE;
+    }
+
+    s->renderer = SDL_CreateRenderer(s->window, nullptr);
+    if (!s->renderer) {
+        SDL_Log("SDL_CreateRenderer failed: %s", SDL_GetError());
+        SDL_DestroyWindow(s->window);
+        delete s;
+        return SDL_APP_FAILURE;
+    }
+
+    return SDL_APP_CONTINUE;
+}
+
+SDL_AppResult SDL_AppEvent(void* /*appstate*/, SDL_Event* event)
+{
+    if (event->type == SDL_EVENT_QUIT)
+        return SDL_APP_SUCCESS;
+    if (event->type == SDL_EVENT_KEY_DOWN && event->key.key == SDLK_ESCAPE)
+        return SDL_APP_SUCCESS;
+    return SDL_APP_CONTINUE;
+}
+
+SDL_AppResult SDL_AppIterate(void* appstate)
+{
+    auto* s = static_cast<AppState*>(appstate);
+
+    SDL_FRect rect;
+
+    SDL_SetRenderDrawColor(s->renderer, 0, 0, 0, 255);
+    SDL_RenderClear(s->renderer);
+
+    SDL_SetRenderDrawColor(s->renderer, 255, 0, 0, 255);
+    rect.x = rect.y = 100;
+    rect.w = 440;
+    rect.h = 280;
+    SDL_RenderFillRect(s->renderer, &rect);
+
+    SDL_RenderPresent(s->renderer);
+
+    return SDL_APP_CONTINUE;
+}
+
+void SDL_AppQuit(void* appstate, SDL_AppResult /*result*/)
+{
+    auto* s = static_cast<AppState*>(appstate);
+    if (!s)
+        return;
+    SDL_DestroyRenderer(s->renderer);
+    SDL_DestroyWindow(s->window);
+    SDL_Quit();
+    delete s;
+}
+}
